Add Makefile support for yaml-test-suite
Add `test-suite` rule to Makefile.  This test clones the libyaml-test
repo (which then clones the libyaml-parser and libyaml-emitter repos)
and calls `make test` in it.  That in turn clone yaml/yaml-test suite.

Normally the subtests will clone the latest libyaml, but this test
passes the current libyaml via an environment variable.

Also add `test-all` to run this together with normal `make test` rule.