Support a default environment
The env files were renamed to be more obvious.

Now pin-* files are symlinks to env-* files.

If a pinning can't be found using the current mainline commit, we
use env/default default which is a symlink to the most recent.

This will probably work most of the time.

A big warning box will be printed showing what happened and where to get
more info.

To try this out:

  git checkout ingy.test-suite-support
  export LIBYAML_TEST_SUITE_RUN_BRANCH=run-test-suite-2
  ln -fs .makefile GNUmakefile
  make distclean
  make test-suite